250908 Blog Module 4 - Week 1/Gradio/Code/requirements.txt
gradio==5.44.1
numpy>=1.23,<3
pillow>=9,<11